+++ src/domain/services/imports/publish-import-to-queue.ts
@@ -1,0 +1,28 @@
+import { publish } from '@/adapters/sqs'
+import type { DetailedUserImport } from '@/domain/entities/import'
+import type { QueueMessage } from '@/domain/entities/queue-message'
+import { config } from '@/env'
+
+export async function publishToQueue(result: DetailedUserImport) {
+  processAndPublish(result.movies, config.sqsQueues.IMPORT_MOVIES_QUEUE)
+  processAndPublish(result.series, config.sqsQueues.IMPORT_SERIES_QUEUE)
+}
+
+async function processAndPublish(
+  items: { id: string; name: string }[],
+  queueUrl: string
+) {
+  if (items.length === 0) return
+
+  const parsedMessages = items.map(item => ({
+    id: item.id,
+    name: item.name,
+  }))
+
+  const message: QueueMessage = {
+    messages: parsedMessages,
+    queueUrl,
+  }
+
+  await publish(message)
+}