+++ src/domain/services/imports/update-import-series-status.ts
@@ -1,6 +1,6 @@
 import { updateImportSeriesStatus as repository } from '@/db/repositories/import-series-repository'
 import { checkAndFinalizeImport } from '@/db/repositories/user-import-repository'
-import type { ImportStatusEnum } from '@/domain/value-objects/import-item-status-enum'
+import type { ImportStatusEnum } from '@/@types/import-item-status-enum'
 
 export type UpdateUserImportInterface = {
   id: string