+++ src/http/controllers/user-import-controller.ts
@@ -6,6 +6,7 @@ import {
 import { providerDispatcher } from '@/domain/dispatchers/import-dispatcher'
 import { GetUserImport } from '@/domain/services/imports/get-user-import-by-id'
 import { DomainError } from '@/domain/errors/domain-error'
+import { publishToQueue } from '@/domain/services/imports/publish-import-to-queue'
 
 const MAXIMUM_FILE_SIZE_IN_BYTES = 1024 * 1024 * 4 // 4mb
 
@@ -31,6 +32,8 @@ export async function createImportController(
     if (result instanceof DomainError) {
       return reply.status(result.status).send({ message: result.message })
     }
+
+    publishToQueue(result)
 
     return reply.status(200).send({ message: 'File processed successfully.' })
   } catch (error) {