+++ src/test/factories/make-import-movies.ts
@@ -1,6 +1,6 @@
 import type { InsertImportMovie } from '@/domain/entities/import-movies'
-import type { ImportStatusEnum } from '@/domain/value-objects/import-item-status-enum'
-import type { UserItemStatus } from '@/domain/value-objects/item-status-enum'
+import type { ImportStatusEnum } from '@/@types/import-item-status-enum'
+import type { UserItemStatus } from '@/@types/item-status-enum'
 
 import { faker } from '@faker-js/faker'