+++ src/adapters/sqs.ts
@@ -1,26 +1,90 @@
+import type { QueueMessage } from '@/domain/entities/queue-message'
+import { config } from '@/env'
 import type { QueueService } from '@/ports/queue-service'
-import { createSqsClient } from '@/sqs'
 import {
   CreateQueueCommand,
   ReceiveMessageCommand,
-  SendMessageCommand,
-  type SQSClient,
+  SendMessageBatchCommand,
+  SQSClient,
 } from '@aws-sdk/client-sqs'
 
-async function createQueue(sqsClient: SQSClient, queueName: string) {
-  const params = { QueueName: queueName }
-  const result = await sqsClient.send(new CreateQueueCommand(params))
-  return result.QueueUrl
+export async function initializeSQS(sqsClient: SQSClient) {
+  if (config.app.APP_ENV === 'production') {
+    return
+  }
+
+  const queues = [
+    config.sqsQueues.IMPORT_MOVIES_QUEUE,
+    config.sqsQueues.IMPORT_SERIES_QUEUE,
+  ]
+
+  for (const queueName of queues) {
+    try {
+      const command = new CreateQueueCommand({ QueueName: queueName })
+
+      const result = await sqsClient.send(command)
+      console.info(`Queue created or exists: ${result.QueueUrl}`)
+    } catch (error) {
+      console.error(`Failed to create queue ${queueName}:`, error)
+      throw error
+    }
+  }
+
+  console.info('All queues have been created.')
 }
 
-async function sendMessage(
-  sqsClient: SQSClient,
-  queueUrl: string,
-  message: string
-) {
-  const params = { QueueUrl: queueUrl, MessageBody: message }
-  const result = await sqsClient.send(new SendMessageCommand(params))
-  return result.MessageId
+export const createSqsClient = () => {
+  return new SQSClient({
+    region: config.sqs.AWS_REGION,
+    endpoint: config.sqs.LOCALSTACK_ENDPOINT || undefined,
+    credentials: {
+      accessKeyId: config.sqs.AWS_ACCESS_KEY_ID,
+      secretAccessKey: config.sqs.AWS_SECRET_ACCESS_KEY,
+    },
+  })
+}
+
+export async function publish({ messages, queueUrl }: QueueMessage) {
+  const sqsClient = createSqsClient()
+  const MAX_BATCH_SIZE = 10 // SQS LIMIT
+
+  const batches = []
+  for (let i = 0; i < messages.length; i += MAX_BATCH_SIZE) {
+    const batch = messages
+      .slice(i, i + MAX_BATCH_SIZE)
+      .map((message, index) => ({
+        Id: `${i + index}`,
+        MessageBody: JSON.stringify(message),
+      }))
+    batches.push(batch)
+  }
+
+  try {
+    for (const batch of batches) {
+      const command = new SendMessageBatchCommand({
+        QueueUrl: queueUrl,
+        Entries: batch,
+      })
+
+      const { Successful = [], Failed = [] } = await sqsClient.send(command)
+
+      if (Successful.length) {
+        console.info(
+          'Batch sent successfully:',
+          Successful.map(msg => msg.Id)
+        )
+      } else {
+        console.warn('No successful messages in batch.')
+      }
+
+      if (Failed.length) {
+        console.error('Failed messages:', Failed)
+      }
+    }
+  } catch (error) {
+    console.error('Error while publishing to SQS:', error)
+    throw error
+  }
 }
 
 async function receiveMessage(sqsClient: SQSClient, queueUrl: string) {
@@ -30,9 +94,7 @@ async function receiveMessage(sqsClient: SQSClient, queueUrl: string) {
 }
 
 const SqsAdapter: QueueService = {
-  createQueue: queueName => createQueue(createSqsClient(), queueName),
-  sendMessage: (queueUrl, message) =>
-    sendMessage(createSqsClient(), queueUrl, message),
+  publish: queueMessage => publish(queueMessage),
   receiveMessage: queueUrl => receiveMessage(createSqsClient(), queueUrl),
 }