+++ src/test/factories/make-import-movies.ts
@@ -1,6 +1,6 @@
 import type { InsertImportMovie } from '@/domain/entities/import-movies'
-import type { ImportStatusEnum } from '@/domain/value-objects/import-item-status-enum'
-import type { UserItemStatus } from '@/domain/value-objects/item-status-enum'
+import type { ImportStatusEnum } from '@/@types/import-item-status-enum'
+import type { UserItemStatus } from '@/@types/item-status-enum'
 
 import { faker } from '@faker-js/faker'
 
@@ -11,6 +11,7 @@ export function makeRawImportMovies(overrides: Overrides) {
   return {
     ...params,
     name: faker.book.title(),
+    __metadata: buildMetadata(),
     ...overrides,
   }
 }
@@ -45,5 +46,33 @@ function buildItemType() {
   return {
     userItemStatus,
     importStatus,
+  }
+}
+
+function buildMetadata() {
+  return {
+    series_animedb_id: 34437,
+    series_title: 'Code Geass: Fukkatsu no Lelouch',
+    series_type: 'Movie',
+    series_episodes: 1,
+    my_id: 0,
+    my_watched_episodes: 1,
+    my_start_date: '2022-12-04',
+    my_finish_date: '2022-12-04',
+    my_rated: '',
+    my_score: 9,
+    my_storage: '',
+    my_storage_value: 0,
+    my_status: 'Completed',
+    my_comments: '',
+    my_times_watched: 0,
+    my_rewatch_value: '',
+    my_priority: 'LOW',
+    my_tags: '',
+    my_rewatching: 0,
+    my_rewatching_ep: 0,
+    my_discuss: 1,
+    my_sns: 'default',
+    update_on_import: 0,
   }
 }