+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Send imports to SQS using Localstack for local developments [(#69)](https://github.com/plotwist-app/plotwist-backend/pull/69)
+
 ### 1.5.5
 
 - Remove docker cache from pulumi [(#68)](https://github.com/plotwist-app/plotwist-backend/pull/68)