+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Send imports to SQS using Localstack for local developments [(#69)](https://github.com/plotwist-app/plotwist-backend/pull/69)
+- Add __metadata field to save raw import [(#72)](https://github.com/plotwist-app/plotwist-backend/pull/72)
 
 ### 1.5.5