+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-- Update `get-user-by-username` query tu handle capitalized usernames [#70](https://github.com/plotwist-app/plotwist-backend/pull/70)
+- Update `get-user-by-username` query to handle capitalized usernames [#70](https://github.com/plotwist-app/plotwist-backend/pull/70)
+
+### Added
+
+- Send imports to SQS using Localstack for local developments [(#69)](https://github.com/plotwist-app/plotwist-backend/pull/69)
+- Add \_\_metadata field to save raw import [(#72)](https://github.com/plotwist-app/plotwist-backend/pull/72)
 
 ### 1.5.5